Use `skip_global_compinit=1` on Ubuntu
The system wide config /etc/zsh/zshrc calls compinit by default.
This is the default on Ubuntu. They provide a way to disable this
by setting `skip_global_compinit=1` (this is described in /etc/zsh/zshrc).
Because we create and cache zcompdump ourselves in $XDG_CACHE_HOME/zsh
we do not want it to be created by default as we lose all the
performance optimization we did.